Cleanup code a bit
