+++ plugins/inputs/openweathermap/openweathermap.go
@@ -22,17 +22,9 @@ import (
 //go:embed sample.conf
 var sampleConfig string
 
-const (
-	// https://openweathermap.org/current#severalid
-	// Call for several city IDs
-	// The limit of locations is 20.
-	owmRequestSeveralCityID int = 20
-
-	defaultBaseURL         = "https://api.openweathermap.org/"
-	defaultResponseTimeout = time.Second * 5
-	defaultUnits           = "metric"
-	defaultLang            = "en"
-)
+// https://openweathermap.org/current#severalid
+// Limit for the number of city IDs per request.
+const maxIDsPerBatch int = 20
 
 type OpenWeatherMap struct {
 	AppID           string          `toml:"app_id"`
@@ -44,6 +36,7 @@ type OpenWeatherMap struct {
 	Units           string          `toml:"units"`
 
 	client        *http.Client
+	cityIDBatches []string
 	baseParsedURL *url.URL
 }
 
@@ -51,48 +44,97 @@ func (*OpenWeatherMap) SampleConfig() string {
 	return sampleConfig
 }
 
+func (n *OpenWeatherMap) Init() error {
+	// Set the default for the base-URL if not given
+	if n.BaseURL == "" {
+		n.BaseURL = "https://api.openweathermap.org/"
+	}
+
+	// Check the unit setting
+	switch n.Units {
+	case "imperial", "standard", "metric":
+	case "":
+		n.Units = "metric"
+	default:
+		return fmt.Errorf("unknown units: %s", n.Units)
+	}
+
+	// Check the language setting
+	switch n.Lang {
+	case "ar", "bg", "ca", "cz", "de", "el", "en", "fa", "fi", "fr", "gl",
+		"hr", "hu", "it", "ja", "kr", "la", "lt", "mk", "nl", "pl",
+		"pt", "ro", "ru", "se", "sk", "sl", "es", "tr", "ua", "vi",
+		"zh_cn", "zh_tw":
+		// Do nothing, those are valid
+	case "":
+		n.Lang = "en"
+	default:
+		return fmt.Errorf("unknown language: %s", n.Lang)
+	}
+
+	// Check the properties to fetch
+	if len(n.Fetch) == 0 {
+		n.Fetch = []string{"weather", "forecast"}
+	}
+	for _, fetch := range n.Fetch {
+		switch fetch {
+		case "forecast", "weather":
+			// Do nothing, those are valid
+		default:
+			return fmt.Errorf("unknown property to fetch: %s", fetch)
+		}
+	}
+
+	// Split the city IDs into batches smaller than the maximum size
+	nBatches := len(n.CityID) / maxIDsPerBatch
+	if len(n.CityID)%maxIDsPerBatch != 0 {
+		nBatches++
+	}
+	batches := make([][]string, nBatches)
+	for i, id := range n.CityID {
+		batch := i / maxIDsPerBatch
+		batches[batch] = append(batches[batch], id)
+	}
+	n.cityIDBatches = make([]string, 0, nBatches)
+	for _, batch := range batches {
+		n.cityIDBatches = append(n.cityIDBatches, strings.Join(batch, ","))
+	}
+
+	// Parse the base-URL used later to construct the property API endpoint
+	u, err := url.Parse(n.BaseURL)
+	if err != nil {
+		return err
+	}
+	n.baseParsedURL = u
+
+	// Create an HTTP client to be used in each collection interval
+	n.client = &http.Client{
+		Transport: &http.Transport{},
+		Timeout:   time.Duration(n.ResponseTimeout),
+	}
+
+	return nil
+}
+
 func (n *OpenWeatherMap) Gather(acc telegraf.Accumulator) error {
 	var wg sync.WaitGroup
-	var strs []string
-
 	for _, fetch := range n.Fetch {
-		if fetch == "forecast" {
-			for _, city := range n.CityID {
-				addr := n.formatURL("/data/2.5/forecast", city)
+		switch fetch {
+		case "forecast":
+			for _, cityID := range n.CityID {
 				wg.Add(1)
-				go func() {
+				go func(city string) {
 					defer wg.Done()
-					status, err := n.gatherURL(addr)
-					if err != nil {
-						acc.AddError(err)
-						return
-					}
-
-					gatherForecast(acc, status)
-				}()
+					acc.AddError(n.gatherForecast(acc, city))
+				}(cityID)
 			}
-		} else if fetch == "weather" {
-			j := 0
-			for j < len(n.CityID) {
-				strs = make([]string, 0)
-				for i := 0; j < len(n.CityID) && i < owmRequestSeveralCityID; i++ {
-					strs = append(strs, n.CityID[j])
-					j++
-				}
-				cities := strings.Join(strs, ",")
-
-				addr := n.formatURL("/data/2.5/group", cities)
+		case "weather":
+			for _, cityIDs := range n.cityIDBatches {
 				wg.Add(1)
-				go func() {
+				go func(cities string) {
 					defer wg.Done()
-					status, err := n.gatherURL(addr)
-					if err != nil {
-						acc.AddError(err)
-						return
-					}
-
-					gatherWeather(acc, status)
-				}()
+					acc.AddError(n.gatherWeatherBatch(acc, cities))
+				}(cityIDs)
 			}
 		}
 	}
@@ -101,122 +143,20 @@ func (n *OpenWeatherMap) Gather(acc telegraf.Accumulator) error {
 	return nil
 }
 
-func (n *OpenWeatherMap) createHTTPClient() *http.Client {
-	if n.ResponseTimeout < config.Duration(time.Second) {
-		n.ResponseTimeout = config.Duration(defaultResponseTimeout)
-	}
-
-	client := &http.Client{
-		Transport: &http.Transport{},
-		Timeout:   time.Duration(n.ResponseTimeout),
-	}
-
-	return client
-}
-
-func (n *OpenWeatherMap) gatherURL(addr string) (*Status, error) {
-	resp, err := n.client.Get(addr)
+func (n *OpenWeatherMap) gatherWeatherBatch(acc telegraf.Accumulator, cities string) error {
+	// Query the data and decode the response
+	addr := n.formatURL("/data/2.5/group", cities)
+	buf, err := n.gatherURL(addr)
 	if err != nil {
-		return nil, fmt.Errorf("error making HTTP request to %q: %w", addr, err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("%s returned HTTP status %s", addr, resp.Status)
+		return fmt.Errorf("querying %q failed: %w", addr, err)
 	}
 
-	mediaType, _, err := mime.ParseMediaType(resp.Header.Get("Content-Type"))
-	if err != nil {
-		return nil, err
+	var status Status
+	if err := json.Unmarshal(buf, &status); err != nil {
+		return fmt.Errorf("parsing JSON response failed: %w", err)
 	}
 
-	if mediaType != "application/json" {
-		return nil, fmt.Errorf("%s returned unexpected content type %s", addr, mediaType)
-	}
-
-	return gatherWeatherURL(resp.Body)
-}
-
-type WeatherEntry struct {
-	Dt     int64 `json:"dt"`
-	Clouds struct {
-		All int64 `json:"all"`
-	} `json:"clouds"`
-	Main struct {
-		Humidity int64   `json:"humidity"`
-		Pressure float64 `json:"pressure"`
-		Temp     float64 `json:"temp"`
-		Feels    float64 `json:"feels_like"`
-	} `json:"main"`
-	Rain struct {
-		Rain1 float64 `json:"1h"`
-		Rain3 float64 `json:"3h"`
-	} `json:"rain"`
-	Snow struct {
-		Snow1 float64 `json:"1h"`
-		Snow3 float64 `json:"3h"`
-	} `json:"snow"`
-	Sys struct {
-		Country string `json:"country"`
-		Sunrise int64  `json:"sunrise"`
-		Sunset  int64  `json:"sunset"`
-	} `json:"sys"`
-	Wind struct {
-		Deg   float64 `json:"deg"`
-		Speed float64 `json:"speed"`
-	} `json:"wind"`
-	ID    int64  `json:"id"`
-	Name  string `json:"name"`
-	Coord struct {
-		Lat float64 `json:"lat"`
-		Lon float64 `json:"lon"`
-	} `json:"coord"`
-	Visibility int64 `json:"visibility"`
-	Weather    []struct {
-		ID          int64  `json:"id"`
-		Main        string `json:"main"`
-		Description string `json:"description"`
-		Icon        string `json:"icon"`
-	} `json:"weather"`
-}
-
-type Status struct {
-	City struct {
-		Coord struct {
-			Lat float64 `json:"lat"`
-			Lon float64 `json:"lon"`
-		} `json:"coord"`
-		Country string `json:"country"`
-		ID      int64  `json:"id"`
-		Name    string `json:"name"`
-	} `json:"city"`
-	List []WeatherEntry `json:"list"`
-}
-
-func gatherWeatherURL(r io.Reader) (*Status, error) {
-	dec := json.NewDecoder(r)
-	status := &Status{}
-	if err := dec.Decode(status); err != nil {
-		return nil, fmt.Errorf("error while decoding JSON response: %w", err)
-	}
-	return status, nil
-}
-
-func gatherSnow(e WeatherEntry) float64 {
-	if e.Snow.Snow1 > 0 {
-		return e.Snow.Snow1
-	}
-	return e.Snow.Snow3
-}
-
-func gatherRain(e WeatherEntry) float64 {
-	if e.Rain.Rain1 > 0 {
-		return e.Rain.Rain1
-	}
-	return e.Rain.Rain3
-}
-
-func gatherWeather(acc telegraf.Accumulator, status *Status) {
+	// Construct the metrics
 	for _, e := range status.List {
 		tm := time.Unix(e.Dt, 0)
 
@@ -224,8 +164,8 @@ func gatherWeather(acc telegraf.Accumulator, status *Status) {
 			"cloudiness":   e.Clouds.All,
 			"humidity":     e.Main.Humidity,
 			"pressure":     e.Main.Pressure,
-			"rain":         gatherRain(e),
-			"snow":         gatherSnow(e),
+			"rain":         e.rain(),
+			"snow":         e.snow(),
 			"sunrise":      time.Unix(e.Sys.Sunrise, 0).UnixNano(),
 			"sunset":       time.Unix(e.Sys.Sunset, 0).UnixNano(),
 			"temperature":  e.Main.Temp,
@@ -250,9 +190,24 @@ func gatherWeather(acc telegraf.Accumulator, status *Status) {
 
 		acc.AddFields("weather", fields, tags, tm)
 	}
+
+	return nil
 }
 
-func gatherForecast(acc telegraf.Accumulator, status *Status) {
+func (n *OpenWeatherMap) gatherForecast(acc telegraf.Accumulator, city string) error {
+	// Query the data and decode the response
+	addr := n.formatURL("/data/2.5/forecast", city)
+	buf, err := n.gatherURL(addr)
+	if err != nil {
+		return fmt.Errorf("querying %q failed: %w", addr, err)
+	}
+
+	var status Status
+	if err := json.Unmarshal(buf, &status); err != nil {
+		return fmt.Errorf("parsing JSON response failed: %w", err)
+	}
+
+	// Construct the metric
 	tags := map[string]string{
 		"city_id":  strconv.FormatInt(status.City.ID, 10),
 		"forecast": "*",
@@ -265,8 +220,8 @@ func gatherForecast(acc telegraf.Accumulator, status *Status) {
 			"cloudiness":   e.Clouds.All,
 			"humidity":     e.Main.Humidity,
 			"pressure":     e.Main.Pressure,
-			"rain":         gatherRain(e),
-			"snow":         gatherSnow(e),
+			"rain":         e.rain(),
+			"snow":         e.snow(),
 			"temperature":  e.Main.Temp,
 			"feels_like":   e.Main.Feels,
 			"wind_degrees": e.Wind.Deg,
@@ -280,47 +235,6 @@ func gatherForecast(acc telegraf.Accumulator, status *Status) {
 		}
 		tags["forecast"] = fmt.Sprintf("%dh", (i+1)*3)
 		acc.AddFields("weather", fields, tags, tm)
-	}
-}
-
-func init() {
-	inputs.Add("openweathermap", func() telegraf.Input {
-		tmout := config.Duration(defaultResponseTimeout)
-		return &OpenWeatherMap{
-			ResponseTimeout: tmout,
-			BaseURL:         defaultBaseURL,
-		}
-	})
-}
-
-func (n *OpenWeatherMap) Init() error {
-	var err error
-	n.baseParsedURL, err = url.Parse(n.BaseURL)
-	if err != nil {
-		return err
-	}
-
-	// Create an HTTP client that is re-used for each
-	// collection interval
-	n.client = n.createHTTPClient()
-
-	switch n.Units {
-	case "imperial", "standard", "metric":
-	case "":
-		n.Units = defaultUnits
-	default:
-		return fmt.Errorf("unknown units: %s", n.Units)
-	}
-
-	switch n.Lang {
-	case "ar", "bg", "ca", "cz", "de", "el", "en", "fa", "fi", "fr", "gl",
-		"hr", "hu", "it", "ja", "kr", "la", "lt", "mk", "nl", "pl",
-		"pt", "ro", "ru", "se", "sk", "sl", "es", "tr", "ua", "vi",
-		"zh_cn", "zh_tw":
-	case "":
-		n.Lang = defaultLang
-	default:
-		return fmt.Errorf("unknown language: %s", n.Lang)
 	}
 
 	return nil
@@ -340,4 +254,35 @@ func (n *OpenWeatherMap) formatURL(path string, city string) string {
 	}
 
 	return n.baseParsedURL.ResolveReference(relative).String()
+}
+
+func (n *OpenWeatherMap) gatherURL(addr string) ([]byte, error) {
+	resp, err := n.client.Get(addr)
+	if err != nil {
+		return nil, fmt.Errorf("error making HTTP request to %q: %w", addr, err)
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("%s returned HTTP status %s", addr, resp.Status)
+	}
+
+	mediaType, _, err := mime.ParseMediaType(resp.Header.Get("Content-Type"))
+	if err != nil {
+		return nil, err
+	}
+
+	if mediaType != "application/json" {
+		return nil, fmt.Errorf("%s returned unexpected content type %s", addr, mediaType)
+	}
+
+	return io.ReadAll(resp.Body)
+}
+
+func init() {
+	inputs.Add("openweathermap", func() telegraf.Input {
+		return &OpenWeatherMap{
+			ResponseTimeout: config.Duration(5 * time.Second),
+		}
+	})
 }
